Refactor core modules to DDD naming standards
- Renamed `cognitive` -> `compute`
- Renamed `engines.py` -> `reasoning.py`
- Renamed `topology` -> `workflow`
- Renamed `co_intelligence.py` -> `intervention.py`
- Moved `nodes.py` from `compute` to `workflow` to resolve circular dependency.
- Fixed all imports globally.
- Updated `rebuild.py` lazy imports and namespace map.
- Deleted obsolete tests as per instruction.
