+++ src/coreason_manifest/spec/core/workflow/nodes.py
